allow to buy barne's bomb bag even if you already have one

+++ source/item.cpp
@@ -8,16 +8,16 @@ namespace mod::item
 {
 	u8 itemGetAnimationFlags[10] = 
 	{
-		static_cast<u16>(ItemFlags::Blue_Rupee),
-		static_cast<u16>(ItemFlags::Yellow_Rupee),
-		static_cast<u16>(ItemFlags::Red_Rupee),
-		static_cast<u16>(ItemFlags::Purple_Rupee),
-		static_cast<u16>(ItemFlags::Orange_Rupee),
-		static_cast<u16>(ItemFlags::Seeds_50),
-		static_cast<u16>(ItemFlags::Arrows_30),
-		static_cast<u16>(ItemFlags::Arrows_20),
-		static_cast<u16>(ItemFlags::Arrows_10),
-		static_cast<u16>(ItemFlags::Arrows_1),
+		static_cast<u8>(ItemFlags::Blue_Rupee),
+		static_cast<u8>(ItemFlags::Yellow_Rupee),
+		static_cast<u8>(ItemFlags::Red_Rupee),
+		static_cast<u8>(ItemFlags::Purple_Rupee),
+		static_cast<u8>(ItemFlags::Orange_Rupee),
+		static_cast<u8>(ItemFlags::Seeds_50),
+		static_cast<u8>(ItemFlags::Arrows_30),
+		static_cast<u8>(ItemFlags::Arrows_20),
+		static_cast<u8>(ItemFlags::Arrows_10),
+		static_cast<u8>(ItemFlags::Arrows_1),
 	};
 	
 	u32 getFlags(u8 item, u32 currentPlayerConditions)
